FINALLY GOT THE BACKGROUND!!! :D
So after getting really frustrated for over an hour with it, i finally
decided to take a look at the Math.cos(a) Javadocs and guess what I
found! arg a IS IN FLIPPING RADIANS!!! RADIANS CLAY!!! RADIANS ARE SO
ANNOYING!!!! Anyways all I had to do was add in a converting algorithm
(that's where the Math.PI/180 comes from) to convert it from angles to
radians to get the proper results!
Also I decided to stick with basing the background color of of timer
(which is also our score) because I like the effect of the background
suddenly changing when you get hit!
Anyways you'd better like the background because it took over an hour to
get working properly!